Fix bounty calc and bounty bug
Reverting baseAmount back to rscAmount - platformFee to not break other logic. Created "totalCost" to render the correct frontend when opening a bounty from the paper page.

A separate bug was found when opening a bounty from the +new. bountyAmount that was being passed to the backend was set to "totalAmount" but should be "rscAmount" since the platform fee is taken elsewhere. totalAmount bundles the platform fee into the bounty amount and then a platform fee is taken on top.